new proj - Copy/ai-customer-service-backend/requirements.txt
langchain==0.1.17
langchain-community==0.0.37
langchain-google-genai==0.0.11
google-generativeai==0.4.1
python-dotenv==1.0.0
chromadb==0.4.15
fastapi==0.120.0
uvicorn==0.38.0
PyMuPDF==1.23.8
langchain-core==0.1.52
faster-whisper
sounddevice 
numpy
pyttsx3
python-dotenv
tqdm
edge-tts
pydub
simpleaudio
soundfile
pygame
PyAudio
keyboard
SpeechRecognition
python-multipart